Move device lock out of NodeManager
Another refactor for optional NodeManager. This moves the device lock
out of the NodeManager so we can still enforce a single device even for
fedimint only wallets.
